Move group/roles into its own service
