+++ src/console/dbseed/domain/db-seed.service.spec.ts
@@ -30,6 +30,7 @@ import { ServiceProvider } from '../../../modules/service-provider/domain/servic
 import { GleicheRolleAnKlasseWieSchuleError } from '../../../modules/personenkontext/specification/error/gleiche-rolle-an-klasse-wie-schule.error.js';
 import { PersonenkontextFactory } from '../../../modules/personenkontext/domain/personenkontext.factory.js';
 import { OrganisationRepository } from '../../../modules/organisation/persistence/organisation.repository.js';
+import { KeycloakGroupRoleService } from '../../../modules/keycloak-administration/domain/keycloak-group-role.service.js';
 
 describe('DbSeedService', () => {
     let module: TestingModule;
@@ -94,6 +95,10 @@ describe('DbSeedService', () => {
                 {
                     provide: KeycloakUserService,
                     useValue: createMock<KeycloakUserService>(),
+                },
+                {
+                    provide: KeycloakGroupRoleService,
+                    useValue: createMock<KeycloakGroupRoleService>(),
                 },
             ],
         }).compile();
