+++ src/modules/keycloak-administration/domain/keycloak-user.service.ts
@@ -1,12 +1,7 @@
 import { Mapper } from '@automapper/core';
 import { getMapperToken } from '@automapper/nestjs';
 import { Inject, Injectable } from '@nestjs/common';
-import {
-    GroupRepresentation,
-    KeycloakAdminClient,
-    RoleRepresentation,
-    type UserRepresentation,
-} from '@s3pweb/keycloak-admin-client-cjs';
+import { KeycloakAdminClient, type UserRepresentation } from '@s3pweb/keycloak-admin-client-cjs';
 import { plainToClass } from 'class-transformer';
 import { validate, ValidationError } from 'class-validator';
 
@@ -269,106 +264,5 @@ export class KeycloakUserService {
         }
 
         return { ok: true, value: this.mapper.map(userReprDto, UserRepresentationDto, UserDo) };
-    }
-
-    public async createGroup(group: string): Promise<Result<string, DomainError>> {
-        const kcAdminClientResult: Result<KeycloakAdminClient, DomainError> =
-            await this.kcAdminService.getAuthedKcAdminClient();
-
-        if (!kcAdminClientResult.ok) {
-            return kcAdminClientResult;
-        }
-
-        const groupName: string = group;
-        const [existingGroup]: GroupRepresentation[] = await kcAdminClientResult.value.groups.find({
-            search: groupName,
-        });
-
-        if (existingGroup) {
-            this.logger.info(`Group already exists: ${groupName}`);
-            return { ok: false, error: new KeycloakClientError('Group name already exists') };
-        }
-
-        try {
-            const groupRepresentation: GroupRepresentation = {
-                name: groupName,
-            };
-
-            const response: { id: string } = await kcAdminClientResult.value.groups.create(groupRepresentation);
-            this.logger.info(`keycloack group created:  ${groupName}`);
-            return { ok: true, value: response.id };
-        } catch (err) {
-            this.logger.error(`Could not create group, message: ${JSON.stringify(err)} `);
-            return { ok: false, error: new KeycloakClientError('Could not create group') };
-        }
-    }
-
-    public async createRole(role: string): Promise<Result<string, DomainError>> {
-        const kcAdminClientResult: Result<KeycloakAdminClient, DomainError> =
-            await this.kcAdminService.getAuthedKcAdminClient();
-
-        if (!kcAdminClientResult.ok) {
-            return kcAdminClientResult;
-        }
-
-        try {
-            const roleName: string = role;
-            const existingRole: RoleRepresentation | undefined = await kcAdminClientResult.value.roles.findOneByName({
-                name: roleName,
-            });
-
-            if (existingRole) {
-                this.logger.info(`Role already exists: ${roleName}`);
-                return { ok: false, error: new KeycloakClientError('Role name already exists') };
-            }
-
-            const roleRepresentation: RoleRepresentation = {
-                name: roleName,
-            };
-
-            const response: { roleName: string } = await kcAdminClientResult.value.roles.create(roleRepresentation);
-            this.logger.info(`Keycloak role created: ${roleName}`);
-            return { ok: true, value: response.roleName };
-        } catch (err) {
-            this.logger.error(`Could not create role, message: ${JSON.stringify(err)} `);
-            return { ok: false, error: new KeycloakClientError('Could not create role') };
-        }
-    }
-
-    public async addRoleToGroup(groupId: string, createdRole: string): Promise<Result<boolean, DomainError>> {
-        const kcAdminClientResult: Result<KeycloakAdminClient, DomainError> =
-            await this.kcAdminService.getAuthedKcAdminClient();
-
-        if (!kcAdminClientResult.ok) {
-            return kcAdminClientResult;
-        }
-
-        try {
-            const roleName: string = createdRole;
-            const role: RoleRepresentation | undefined = await kcAdminClientResult.value.roles.findOneByName({
-                name: roleName,
-            });
-
-            if (!role || !role.id || !role.name) {
-                this.logger.error(`Role not found or id/name is undefined for: ${roleName}`);
-                return { ok: false, error: new KeycloakClientError('Role not found or id/name is undefined') };
-            }
-
-            await kcAdminClientResult.value.groups.addRealmRoleMappings({
-                id: groupId,
-                roles: [
-                    {
-                        id: role.id,
-                        name: role.name,
-                    },
-                ],
-            });
-
-            this.logger.info(`Role ${roleName} added to group with ID: ${groupId}`);
-            return { ok: true, value: true };
-        } catch (err) {
-            this.logger.error(`Could not add role to group, message: ${JSON.stringify(err)}`);
-            return { ok: false, error: new KeycloakClientError('Could not add role to group') };
-        }
     }
 }
