+++ src/modules/keycloak-administration/keycloak-administration.module.ts
@@ -6,10 +6,17 @@ import { UserMapperProfile } from './domain/keycloak-client/user.mapper.profile.
 import { KeycloakUserService } from './domain/keycloak-user.service.js';
 import { LoggerModule } from '../../core/logging/logger.module.js';
 import { KeycloakConfigModule } from './keycloak-config.module.js';
+import { KeycloakGroupRoleService } from './domain/keycloak-group-role.service.js';
 
 @Module({
     imports: [LoggerModule.register(KeycloakAdministrationModule.name), KeycloakConfigModule],
-    providers: [UserMapperProfile, KeycloakAdminClient, KeycloakUserService, KeycloakAdministrationService],
-    exports: [KeycloakUserService],
+    providers: [
+        UserMapperProfile,
+        KeycloakAdminClient,
+        KeycloakUserService,
+        KeycloakGroupRoleService,
+        KeycloakAdministrationService,
+    ],
+    exports: [KeycloakUserService, KeycloakGroupRoleService],
 })
 export class KeycloakAdministrationModule {}
